Chit Billing

+++ SS SOFTWARE CHIT/FRM_WHATSAPP_SETTINGS.cs
@@ -19,7 +19,7 @@ namespace SS_SOFTWARE_CHIT
         OleDbConnection con = new OleDbConnection();
         OleDbCommand cmd = new OleDbCommand();
         WhatsApp app;
-            
+
         public FRM_WHATSAPP_SETTINGS(WhatsApp whatsappInitialize)
         {
             InitializeComponent();
@@ -234,20 +234,28 @@ namespace SS_SOFTWARE_CHIT
         {
             if (e.KeyCode == Keys.Escape)
             {
-                if (MessageBox.Show("DO YOU WANT TO EXIT?", "SS SOFTWARE", MessageBoxButtons.YesNo, MessageBoxIcon.Information) == DialogResult.Yes)
+                if (pnlmain.Visible == true)
                 {
-                    this.Hide();
-                    FRM_SETTINGS Settings = new FRM_SETTINGS(app);
-                    Settings.Show();
+                    if (MessageBox.Show("DO YOU WANT TO EXIT?", "SS SOFTWARE", MessageBoxButtons.YesNo, MessageBoxIcon.Information) == DialogResult.Yes)
+                    {
+                        this.Hide();
+                        FRM_SETTINGS Settings = new FRM_SETTINGS(app);
+                        Settings.Show();
+                    }
+                }
+                else
+                {
+                    pnlmsg.Hide();
+                    pnlmain.Show();
                 }
             }
         }
 
         private void btnbold_Click(object sender, EventArgs e)
         {
-            if(txtmsg.SelectionLength > 1)
+            if (txtmsg.SelectionLength > 1)
             {
-                txtmsg.SelectedText = "*" + txtmsg.SelectedText + "*";
+                txtmsg.SelectedText = "%20" + "*" + txtmsg.SelectedText + "*" + "%20";
             }
             else
             {
@@ -260,7 +268,7 @@ namespace SS_SOFTWARE_CHIT
         {
             if (txtmsg.SelectionLength > 1)
             {
-                txtmsg.SelectedText = "_" + txtmsg.SelectedText + "_";
+                txtmsg.SelectedText = "%20" + "_" + txtmsg.SelectedText + "_" + "%20";
             }
             else
             {
@@ -274,13 +282,13 @@ namespace SS_SOFTWARE_CHIT
         {
             string msg = txtmsg.Text;
             string dateRemove = "{DATE}";
-            if(chkdate.Checked==true)
+            if (chkdate.Checked == true)
             {
                 txtmsg.Text = txtmsg.Text + "{DATE}";
             }
             else
             {
-                msg = msg.Replace(dateRemove,string.Empty);
+                msg = msg.Replace(dateRemove, string.Empty);
                 txtmsg.Text = msg;
             }
             txtmsg.Focus();
@@ -373,21 +381,33 @@ namespace SS_SOFTWARE_CHIT
 
         private void btnsave_Click(object sender, EventArgs e)
         {
-            if (MessageBox.Show("DO YOU WANT TO SAVE???", "SS SOFTWARE", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+            if (txtmsg.Text.Contains(" "))
             {
-                con = new OleDbConnection(path);
-                con.Open();
-                cmd.Connection = con;
-                cmd.CommandText = "UPDATE Whatsapp_db set f_msg='" + txtmsg.Text + "' , f_customer_id ='"+chkcustomerid.Checked+ "' , f_customer_name ='" + chkcustomername.Checked + "', f_mobile_no ='" + chkmobileno.Checked + "', f_type ='" + chktype.Checked + "',f_amount ='" + chkamount.Checked + "', f_date ='" + chkdate.Checked + "'";
-                cmd.ExecuteNonQuery();
-                con.Close();
-                string SP = Application.StartupPath + "\\BACKUP\\";
-                string ThisDB = Application.StartupPath + "\\DATABASE\\Settings_db.accdb";
-                string Destitnation = SP + "\\Settings_db " + DateTime.Now.ToString(" dd-MM-yyyy hh-mm-ss") + ".bak";
-                File.Copy(ThisDB, Destitnation);
-                pnlmain.Visible = true;
-                MessageBox.Show("UPDATED SUCCESSFULLY!", "SS SOFTWARE", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                txtmsg.Text = txtmsg.Text.Replace(" ", "%20");
             }
+            else
+            {
+                if (MessageBox.Show("DO YOU WANT TO SAVE???", "SS SOFTWARE", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+                {
+                    con = new OleDbConnection(path);
+                    con.Open();
+                    cmd.Connection = con;
+                    cmd.CommandText = "UPDATE Whatsapp_db set f_msg='" + txtmsg.Text + "' , f_customer_id ='" + chkcustomerid.Checked + "' , f_customer_name ='" + chkcustomername.Checked + "', f_mobile_no ='" + chkmobileno.Checked + "', f_type ='" + chktype.Checked + "',f_amount ='" + chkamount.Checked + "', f_date ='" + chkdate.Checked + "'";
+                    cmd.ExecuteNonQuery();
+                    con.Close();
+                    string SP = Application.StartupPath + "\\BACKUP\\";
+                    string ThisDB = Application.StartupPath + "\\DATABASE\\Settings_db.accdb";
+                    string Destitnation = SP + "\\Settings_db " + DateTime.Now.ToString(" dd-MM-yyyy hh-mm-ss") + ".bak";
+                    File.Copy(ThisDB, Destitnation);
+                    MessageBox.Show("UPDATED SUCCESSFULLY!", "SS SOFTWARE", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                }
+            }
+        }
+
+        private void btnmsg_Click(object sender, EventArgs e)
+        {
+            pnlmsg.Show();
+            pnlmain.Hide();
         }
     }
 }
